+++ src/DataSources/Pdf/Chunker/IPdfChunker.cs
@@ -1,0 +1,12 @@
+﻿using SimpleRag.DataProviders.Models;
+using System.Diagnostics.Metrics;
+using System.Text;
+using UglyToad.PdfPig;
+using UglyToad.PdfPig.Content;
+
+namespace SimpleRag.DataSources.Pdf.Chunker;
+
+public interface IPdfChunker
+{
+    PdfChunk[] GetChunks(FileContent file);
+}