+++ src/Integrations/GitHub/Models/GitHubRepository.cs
@@ -3,15 +3,4 @@
 /// <summary>
 /// Represent a GitHub Repository
 /// </summary>
-public class GitHubRepository
-{
-    /// <summary>
-    /// Gets or sets the repository owner.
-    /// </summary>
-    public required string? Owner { get; init; }
-
-    /// <summary>
-    /// Gets or sets the repository name.
-    /// </summary>
-    public required string? Name { get; init; }
-}
+public record GitHubRepository(string Owner, string Name);