+++ src/ServiceCollectionExtensions.cs
@@ -4,6 +4,7 @@ using SimpleRag.DataSources.CSharp;
 using SimpleRag.DataSources.CSharp.Chunker;
 using SimpleRag.DataSources.Markdown;
 using SimpleRag.DataSources.Markdown.Chunker;
+using SimpleRag.DataSources.Pdf.Chunker;
 using SimpleRag.Integrations.GitHub;
 using SimpleRag.Integrations.GitHub.Models;
 using SimpleRag.VectorStorage;
@@ -25,17 +26,19 @@ public static class ServiceCollectionExtensions
         VectorStoreConfiguration configuration,
         Func<IServiceProvider, T>? vectorStoreFactory) where T : Microsoft.Extensions.VectorData.VectorStore
     {
-        services.AddSimpleRagWithGitHubIntegration(configuration, vectorStoreFactory, string.Empty);
-    }
+        services.AddScoped<IVectorStoreQuery, VectorStoreQuery>();
+        services.AddScoped<IVectorStoreCommand, VectorStoreCommand>();
+        services.AddSingleton(configuration);
+        if (vectorStoreFactory != null)
+        {
+            services.AddScoped<Microsoft.Extensions.VectorData.VectorStore, T>(vectorStoreFactory);
+        }
 
-    /// <summary>
-    /// Registers SimpleRag services including GitHub integration.
-    /// </summary>
-    [PublicAPI]
-    public static void AddSimpleRagWithGitHubIntegration<T>(
-        this IServiceCollection services,
-        VectorStoreConfiguration configuration,
-        Func<IServiceProvider, T>? vectorStoreFactory, string githubPatToken) where T : Microsoft.Extensions.VectorData.VectorStore
-    {
+        services.AddScoped<ICSharpChunker, CSharpChunker>();
+        services.AddScoped<IMarkdownChunker, MarkdownChunker>();
+        services.AddScoped<IPdfChunker, PdfChunker>();
+        services.AddScoped<Ingestion>();
+        services.AddScoped<Search>();
+        services.AddScoped<DataManagement>();
     }
 }